+++ internal/messenger/selectors.go
@@ -3,8 +3,11 @@ package messenger
 const (
 	feedCardSelector      = "div.a11y_feed_card_item"
 	feedTitleSelector     = "div._2b17aec6"
+	feedPreviewSelector   = "span._5d806a78"
+	feedTimeSelector      = "span._49737e67"
 	recentScrollSelector  = "div.simplebar-content-wrapper"
 	searchTriggerSelector = ".appNavbar-search-input-wrapper"
 	chatTitleSelector     = "#root-messenger-chat span.chatWindow_chatName"
 	rightPanelSelector    = "#root-messenger-chat"
+	messageItemSelector   = "#root-messenger-chat .js-message-item"
 )